+++ src/app/page.tsx
@@ -5,7 +5,7 @@ export default function Home() {
     return (
         <main className="flex min-h-screen flex-col items-center justify-between p-24">
             <p>This is main</p>
-            <Link href="/quiz">go to Quiz</Link>
+            <Link href="/quiz/0">go to Quiz</Link>
         </main>
     )
 }
