+++ src/app/quiz/[id]/one-light.css
